Updates - rename project; add new Bicep and YAML files
