+++ AzureFunctionsDemo/.gitignore
