+++ AzureFunctionsDemo/GlobalUsings.cs
@@ -1,0 +1,17 @@
+﻿global using AzureFunctionsDotNet6Demo;
+global using AzureFunctionsDotNet6Demo.Models;
+global using AzureFunctionsDotNet6Demo.Services;
+global using Microsoft.AspNetCore.Http;
+global using Microsoft.AspNetCore.Mvc;
+global using Microsoft.Azure.Documents;
+global using Microsoft.Azure.Documents.Client;
+global using Microsoft.Azure.Functions.Extensions.DependencyInjection;
+global using Microsoft.Azure.WebJobs;
+global using Microsoft.Azure.WebJobs.Extensions.Http;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Logging;
+global using Newtonsoft.Json;
+global using System;
+global using System.Collections.Generic;
+global using System.IO;
+global using System.Threading.Tasks;
