+++ AzureFunctionsDemo/Properties/ServiceDependencies/cmh-cosmos-demo - Zip Deploy/profile.arm.json
