+++ AzureFunctionsDemo/Properties/ServiceDependencies/cmh-func-demo - Zip Deploy/profile.arm.json
