+++ AzureFunctionsDemo/Services/IInventoryUpdateService.cs
@@ -1,0 +1,5 @@
+﻿namespace AzureFunctionsDotNet6Demo.Services;
+public interface IInventoryUpdateService
+{
+    Task UpdateAvailableInventoryAsync(IDocumentClient documentClient, OrderLineItem lineItem);
+}
