--- AzureFunctionsDotNet6Demo.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32630.192
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureFunctionsDotNet6Demo", "AzureFunctionsDotNet6Demo\AzureFunctionsDotNet6Demo.csproj", "{0CE147CD-4F9A-4B72-8979-470B31CAC8A2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0CE147CD-4F9A-4B72-8979-470B31CAC8A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0CE147CD-4F9A-4B72-8979-470B31CAC8A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0CE147CD-4F9A-4B72-8979-470B31CAC8A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0CE147CD-4F9A-4B72-8979-470B31CAC8A2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7121DEB7-BB84-4512-B41C-7EF19A74F67D}
-	EndGlobalSection
-EndGlobal
