--- AzureFunctionsDotNet6Demo/Models/Order.cs
@@ -1,31 +1,0 @@
-﻿using System.Collections.Generic;
-using Microsoft.Azure.Documents;
-using Newtonsoft.Json;
-
-namespace AzureFunctionsDotNet6Demo.Models
-{
-    public class Order
-    {
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("customerNumber")]
-        public string CustomerNumber { get; set; }
-
-        [JsonProperty("lineItems")]
-        public IList<OrderLineItem> LineItems { get; set; }
-
-        public static Order FromDocument(Document doc)
-        {
-            var orderString = JsonConvert.SerializeObject(doc);
-            var order = JsonConvert.DeserializeObject<Order>(orderString);
-
-            return order;
-        }
-
-        public static Order FromString(string orderJson)
-        {
-            return JsonConvert.DeserializeObject<Order>(orderJson);
-        }
-    }
-}
