--- AzureFunctionsDotNet6Demo/Models/OrderLineItem.cs
@@ -1,16 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace AzureFunctionsDotNet6Demo.Models
-{
-    public class OrderLineItem
-    {
-        [JsonProperty("productId")]
-        public string ProductId { get; set; }
-
-        [JsonProperty("quantity")]
-        public int Quantity { get; set; }
-
-        [JsonProperty("category")]
-        public string Category { get; set; }
-    }
-}
