+++ DeployToAzure/DeployInfrastructureEnvironments.yml
@@ -1,0 +1,92 @@
+﻿# --------------------------------------------------------------------------------
+# Pipeline to deploy Azure Resources for All Environments for the Azure Function Example Project
+# --------------------------------------------------------------------------------
+# TODO: create variable group and use those values
+# --------------------------------------------------------------------------------
+
+trigger: none # fire this manually for now
+# trigger:
+# - main
+
+pool:
+  vmImage: ubuntu-latest
+
+name: $(date:yyyy).$(date:MM).$(date:dd)$(rev:.r)
+
+# --------------------------------------------------------------------------------
+stages:
+- stage: DeployDEVResources
+  displayName: DEV Deploy Resources
+  variables:
+  - group: 'FunctionDemoDevVariables'
+  jobs:
+  - template: ado_deploy_environment.yml
+    parameters:
+      environmentCode: $(environmentCode)
+      environmentUpper: $(environmentUpper)
+      resourceGroupName: $(resourceGroupName)
+      subscriptionName: $(subscriptionName)
+      subscriptionId: $(subscriptionId)
+      regionName: $(regionName)
+      orgPrefix: $(orgPrefix)
+      appPrefix: $(appPrefix)
+      appSuffix: $(appSuffix)
+      storageSku: $(storageSku)
+      functionAppSku: $(functionAppSku)
+      functionAppSkuFamily: $(functionAppSkuFamily)
+      functionAppSkuTier: $(functionAppSkuTier)
+      runDateTime: 'Pipeline-$(Build.BuildNumber)'
+    # parameters:
+    #   environmentCode: 'dev'
+    #   environmentUpper: 'DEV'
+    #   resourceGroupName: 'rg_functionexample'
+    #   subscriptionName: 'Lyles Azure Sandbox (6d1ced742-2c89-420b-a12a-6d9dc6d48c43)'
+    #   subscriptionId: 'd1ced742-2c89-420b-a12a-6d9dc6d48c43'
+    #   regionName: 'eastus'
+    #   orgPrefix: 'lll'
+    #   appPrefix: 'funcdemo'
+    #   appSuffix: '1'
+    #   storageSku: 'Standard_LRS'
+    #   functionAppSku: 'Y1'
+    #   functionAppSkuFamily: 'Y'
+    #   functionAppSkuTier: 'Dynamic'
+    #   runDateTime: 'Pipeline-$(Build.BuildNumber)'
+
+- stage: DeployQAResources
+  displayName: QA Deploy Resources
+  dependsOn: DeployDEVResources
+  condition: succeeded('DeployDEVResources')
+  variables:
+  - group: 'FunctionDemoQAVariables'
+  jobs:
+  - template: ado_deploy_environment.yml
+    parameters:
+      environmentCode: $(environmentCode)
+      environmentUpper: $(environmentUpper)
+      resourceGroupName: $(resourceGroupName)
+      subscriptionName: $(subscriptionName)
+      subscriptionId: $(subscriptionId)
+      regionName: $(regionName)
+      orgPrefix: $(orgPrefix)
+      appPrefix: $(appPrefix)
+      appSuffix: $(appSuffix)
+      storageSku: $(storageSku)
+      functionAppSku: $(functionAppSku)
+      functionAppSkuFamily: $(functionAppSkuFamily)
+      functionAppSkuTier: $(functionAppSkuTier)
+      runDateTime: 'Pipeline-$(Build.BuildNumber)'
+    # parameters:
+    #   environmentCode: 'qa'
+    #   environmentUpper: 'QA'
+    #   resourceGroupName: 'rg_functionexample-qa'
+    #   subscriptionName: 'Lyles Azure Sandbox (6d1ced742-2c89-420b-a12a-6d9dc6d48c43)'
+    #   subscriptionId: 'd1ced742-2c89-420b-a12a-6d9dc6d48c43'
+    #   regionName: 'eastus'
+    #   orgPrefix: 'lll'
+    #   appPrefix: 'funcdemo'
+    #   appSuffix: '1'
+    #   storageSku: 'Standard_LRS'
+    #   functionAppSku: 'Y1'
+    #   functionAppSkuFamily: 'Y'
+    #   functionAppSkuTier: 'Dynamic'
+    #   runDateTime: 'Pipeline-$(Build.BuildNumber)'
