+++ ch-deploy/infrastructure/build.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+bicep build cosmosfunctionsdemo.bicep
